:bug: fix : AI 코드 리뷰 반영

+++ src/main/java/io/mallang/order/domain/OrderItemCommand.java
@@ -1,4 +1,6 @@
 package io.mallang.order.domain;
 
-public record OrderItemCommand(String productId, int quantity, int price) {
+import java.math.BigDecimal;
+
+public record OrderItemCommand(String productId, int quantity, BigDecimal price) {
 }
